extract OnlineTrackerListener from VrpOptimizerWithOnlineTracking, delete VrpOptimizerWithOnlineTracking
