+++ contribs/drt/src/main/java/org/matsim/contrib/drt/optimizer/DrtOptimizer.java
@@ -19,11 +19,11 @@
 
 package org.matsim.contrib.drt.optimizer;
 
-import org.matsim.contrib.dvrp.optimizer.VrpOptimizerWithOnlineTracking;
+import org.matsim.contrib.dvrp.optimizer.VrpOptimizer;
 import org.matsim.core.mobsim.framework.listeners.MobsimBeforeSimStepListener;
 
 /**
  * @author michalm
  */
-public interface DrtOptimizer extends VrpOptimizerWithOnlineTracking, MobsimBeforeSimStepListener {
+public interface DrtOptimizer extends VrpOptimizer, MobsimBeforeSimStepListener {
 }
