+++ contribs/drt/src/main/java/org/matsim/contrib/drt/run/DrtModeQSimModule.java
@@ -160,8 +160,7 @@ public class DrtModeQSimModule extends AbstractDvrpModeQSimModule {
 
 		bindModal(VrpAgentLogic.DynActionCreator.class).
 				toProvider(modalProvider(getter -> new DrtActionCreator(getter.getModal(PassengerEngine.class),
-						getter.getModal(DrtOptimizer.class), getter.get(MobsimTimer.class),
-						getter.get(DvrpConfigGroup.class)))).
+						getter.get(MobsimTimer.class), getter.get(DvrpConfigGroup.class)))).
 				asEagerSingleton();
 
 		bindModal(PassengerRequestCreator.class).toProvider(new Provider<DrtRequestCreator>() {
