+++ contribs/taxi/src/main/java/org/matsim/contrib/taxi/optimizer/TaxiOptimizer.java
@@ -19,11 +19,12 @@
 
 package org.matsim.contrib.taxi.optimizer;
 
-import org.matsim.contrib.dvrp.optimizer.VrpOptimizerWithOnlineTracking;
+import org.matsim.contrib.dvrp.optimizer.VrpOptimizer;
+import org.matsim.contrib.dvrp.tracker.OnlineTrackerListener;
 import org.matsim.core.mobsim.framework.listeners.MobsimBeforeSimStepListener;
 
 /**
  * @author michalm
  */
-public interface TaxiOptimizer extends VrpOptimizerWithOnlineTracking, MobsimBeforeSimStepListener {
+public interface TaxiOptimizer extends VrpOptimizer, OnlineTrackerListener, MobsimBeforeSimStepListener {
 }
